+++ lib/src/server.dart
@@ -75,7 +75,16 @@ abstract class Server {
         _process = process;
         _process
           ..stdout.listen((stdOut) => _log.finer(UTF8.decoder.convert(stdOut)))
-          ..stderr.listen((stdErr) => _log.severe(UTF8.decoder.convert(stdErr)))
+          ..stderr.listen((stdErr) {
+              final text = UTF8.decoder.convert(stdErr);
+              if(text.startsWith('WARN: ')) {
+                _log.warning(text);
+              } else if(text.startsWith('ERROR: ')) {
+                _log.severe(text);
+              } else {
+                _log.finer(text);
+              }
+          })
           ..exitCode.then((exitCode) {
             _process = null;
             _exitCode = exitCode;